Use `size_t` instead of `uint32_t`

+++ cpp/Platform.Hashing/Platform.Hashing.Combine.h
@@ -12,7 +12,7 @@ namespace Platform::Hashing
     }
 
     template<typename T>
-    void Combine(std::uint32_t& hashAccumulator, const T* value, std::size_t length)
+    void Combine(std::size_t& hashAccumulator, const T* value, std::size_t length)
     {
         hashAccumulator = Internal::crc32(reinterpret_cast<const std::uint8_t*>(value), length * sizeof(T), hashAccumulator);
     }
